+++ lib/src/chart/line_chart/line_chart_data.dart
@@ -1569,6 +1569,12 @@ class LineTouchTooltipData with EquatableMixin {
   /// The tooltip background color.
   final Color tooltipBgColor;
 
+  /// The tooltip border color.
+  final Color tooltipBorderColor;
+
+  /// The tooltip border stroke width.
+  final double tooltipBorderWidth;
+
   /// Sets a rounded radius for the tooltip.
   final double tooltipRoundedRadius;
 
@@ -1611,6 +1617,8 @@ class LineTouchTooltipData with EquatableMixin {
   /// also you can set [fitInsideVertically] true to force it to shift inside the chart vertically.
   LineTouchTooltipData({
     Color? tooltipBgColor,
+    Color? tooltipBorderColor,
+    double? tooltipBorderWidth,
     double? tooltipRoundedRadius,
     EdgeInsets? tooltipPadding,
     double? tooltipMargin,
@@ -1621,6 +1629,8 @@ class LineTouchTooltipData with EquatableMixin {
     bool? showOnTopOfTheChartBoxArea,
     double? rotateAngle,
   })  : tooltipBgColor = tooltipBgColor ?? Colors.blueGrey.darken(15),
+        tooltipBorderColor = tooltipBorderColor ?? Colors.black,
+        tooltipBorderWidth = tooltipBorderWidth ?? 0.0,
         tooltipRoundedRadius = tooltipRoundedRadius ?? 4,
         tooltipPadding = tooltipPadding ??
             const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
@@ -1637,6 +1647,8 @@ class LineTouchTooltipData with EquatableMixin {
   @override
   List<Object?> get props => [
         tooltipBgColor,
+        tooltipBorderColor,
+        tooltipBorderWidth,
         tooltipRoundedRadius,
         tooltipPadding,
         tooltipMargin,