+++ lib/src/chart/line_chart/line_chart_painter.dart
@@ -23,6 +23,7 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
       _extraLinesPaint,
       _touchLinePaint,
       _bgTouchTooltipPaint,
+      _bgTouchTooltipBorderPaint,
       _imagePaint;
 
   /// Paints [data] into canvas, it is the animating [LineChartData],
@@ -53,6 +54,10 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
 
     _bgTouchTooltipPaint = Paint()
       ..style = PaintingStyle.fill
+      ..color = Colors.white;
+
+    _bgTouchTooltipBorderPaint = Paint()
+      ..style = PaintingStyle.stroke
       ..color = Colors.white;
 
     _imagePaint = Paint();
@@ -1357,6 +1362,10 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
         bottomLeft: radius,
         bottomRight: radius);
     _bgTouchTooltipPaint.color = tooltipData.tooltipBgColor;
+    if (tooltipData.tooltipBorderWidth > 0) {
+      _bgTouchTooltipBorderPaint.color = tooltipData.tooltipBorderColor;
+      _bgTouchTooltipBorderPaint.strokeWidth = tooltipData.tooltipBorderWidth;
+    }
 
     final rotateAngle = tooltipData.rotateAngle;
     final rectRotationOffset =
@@ -1373,6 +1382,9 @@ class LineChartPainter extends AxisChartPainter<LineChartData> {
       angle: rotateAngle,
       drawCallback: () {
         canvasWrapper.drawRRect(roundedRect, _bgTouchTooltipPaint);
+        if (tooltipData.tooltipBorderWidth > 0) {
+          canvasWrapper.drawRRect(roundedRect, _bgTouchTooltipBorderPaint);
+        }
       },
     );