+++ lib/src/chart/scatter_chart/scatter_chart_data.dart
@@ -379,6 +379,12 @@ class ScatterTouchTooltipData with EquatableMixin {
   /// The tooltip background color.
   final Color tooltipBgColor;
 
+  /// The tooltip border color.
+  final Color tooltipBorderColor;
+
+  /// The tooltip border stroke width.
+  final double tooltipBorderWidth;  
+
   /// Sets a rounded radius for the tooltip.
   final double tooltipRoundedRadius;
 
@@ -414,6 +420,8 @@ class ScatterTouchTooltipData with EquatableMixin {
   /// also you can set [fitInsideVertically] true to force it to shift inside the chart vertically.
   ScatterTouchTooltipData({
     Color? tooltipBgColor,
+    Color? tooltipBorderColor,
+    double? tooltipBorderWidth,
     double? tooltipRoundedRadius,
     EdgeInsets? tooltipPadding,
     double? maxContentWidth,
@@ -422,6 +430,8 @@ class ScatterTouchTooltipData with EquatableMixin {
     bool? fitInsideVertically,
     double? rotateAngle,
   })  : tooltipBgColor = tooltipBgColor ?? Colors.blueGrey.darken(15),
+        tooltipBorderColor = tooltipBorderColor ?? Colors.black,
+        tooltipBorderWidth = tooltipBorderWidth ?? 0.0,
         tooltipRoundedRadius = tooltipRoundedRadius ?? 4,
         tooltipPadding = tooltipPadding ??
             const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
@@ -436,6 +446,8 @@ class ScatterTouchTooltipData with EquatableMixin {
   @override
   List<Object?> get props => [
         tooltipBgColor,
+        tooltipBorderColor,
+        tooltipBorderWidth,
         tooltipRoundedRadius,
         tooltipPadding,
         maxContentWidth,
@@ -449,6 +461,8 @@ class ScatterTouchTooltipData with EquatableMixin {
   /// and replaces provided values.
   ScatterTouchTooltipData copyWith({
     Color? tooltipBgColor,
+    Color? tooltipBorderColor,
+    double? tooltipBorderWidth,
     double? tooltipRoundedRadius,
     EdgeInsets? tooltipPadding,
     double? maxContentWidth,
@@ -459,6 +473,8 @@ class ScatterTouchTooltipData with EquatableMixin {
   }) {
     return ScatterTouchTooltipData(
       tooltipBgColor: tooltipBgColor ?? this.tooltipBgColor,
+      tooltipBorderColor: tooltipBorderColor ?? this.tooltipBorderColor,
+      tooltipBorderWidth: tooltipBorderWidth ?? this.tooltipBorderWidth,
       tooltipRoundedRadius: tooltipRoundedRadius ?? this.tooltipRoundedRadius,
       tooltipPadding: tooltipPadding ?? this.tooltipPadding,
       maxContentWidth: maxContentWidth ?? this.maxContentWidth,